requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
elastic-transport==8.4.0
elasticsearch==8.5.3
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.2
python-dotenv==0.21.0
requests==2.28.1
typing-extensions==4.4.0
urllib3==1.26.13
